Added Support for partial module import
