+++ types/redact-axios-error/axios-error-interceptor.d.ts
@@ -1,0 +1,4 @@
+import { AxiosErrorGroomer } from './axios-error-groomer';
+
+export function getAxiosErrorInterceptor(axiosErrorGroomer?: AxiosErrorGroomer | null): ((error: any) => Promise<never>);
+export const config: AxiosErrorGroomer | null;
